Extend ClankStartupOptimizations to WebView/WebLayer in testing_config
This experiment has a potential impact on WebView/WebLayer, as tasks
of BrowserTaskType::kBootstrap are run as part of //content startup
(specifically, BrowserMainLoop::CreateStartupTasks()). This //content
startup code is invoked by these platforms (via their calls to
BrowserMainRunner::Initialize()).

As we are extending the experiment to these platforms, we add a config
to fieldtrial_testing_config.json. Note that the only feature relevant
to these platforms is TreatBootstrapAsDefault; the other two features
are defined in //chrome and impact the behavior of Chrome only. However,
we give the experiment same configuration as the currently-running
ClankStartupOptimizations experiment as we will fold this experiment
into the main experiment.  Specifically, once we have analyzed the
results on beta, we will extend the main experiment
(currently running on stable) to these platforms.

Bug: 1258621
Change-Id: I7ee12c046ebd705d16379ebae9f9a5bc9d740b90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3484375
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974573}
NOKEYCHECK=True
GitOrigin-RevId: 8c6aeec9ccc4501b296abd48d20cfe84fe5bda20