subframe-reuse: remove unwanted experiment groups
This change removes all experiment groups but kConstant from the
SubframeShutdownDelay experiment, and removes it from Linux.

The experiment arms have not proven themselves better than kConstant
(which has the benefit of being simpler), and the experiment has not
performed well on Linux. A separate g3 change will be made to disable
Linux and these arms on Finch, but this has to be done first as the
GCL config requires that it match fieldtrial_testing_config.json.

The code associated with the removed experiment arms will be removed
in a separate change.

Bug: 894253
Change-Id: I8b4d1e8d5b36a22651ff29d30b152001e4df02ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3533405
Reviewed-by: David Bienvenu <davidbienvenu@chromium.org>
Commit-Queue: Jesse McKenna <jessemckenna@google.com>
Cr-Commit-Position: refs/heads/main@{#982458}
NOKEYCHECK=True
GitOrigin-RevId: 6dc8d78e817d6cdc14587b50169ae559f6d27b23