[PreconnectPriorities] Add fieldtrial_testing_config entry
This CL adds a fieldtrial_testing_config.py entry for the
PreconnectPriorities experiment, which is an augmented version of the
previously-running TreatPreconnectAsDefault experiment. We changed the
entry to turn on giving the preconnect task queue highest priority
rather than putting preconnect tasks in the default queue, as the
former seems more likely to ship.

Bug: 1257582
Change-Id: I254197531691d2e5d765c8c953ceb755a5b57c3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3550267
Reviewed-by: Egor Pasko <pasko@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#985262}
NOKEYCHECK=True
GitOrigin-RevId: ac10bbe97dded19af1cfa5756e185e348285fdc3