compositor: fix setting priority_cutoff_when_visible
The value was getting clobbered later on in the function. Added
test coverage too.

This also updates the testing config as the data indicated
release_tile_resources_for_hidden_layers was negative. But, based on
this bug, the data is completely wrong!

BUG=1289362
TEST=covered by tests

Change-Id: I43555ca73a3672c6fac095fda99d7affce32be45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3552552
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#986046}
NOKEYCHECK=True
GitOrigin-RevId: e7153e68ef5b12498e2a69eebc3c3c13351b2d7a