[reland] [privacy sandbox settings] Add dialog pixel tests
Add pixel tests for consent and notice privacy sandbox dialogs. More
tests in this suite will be added in a separate CLs (testing different
dialog heights, expanded section etc).

This is a reland of this CL:
https://chromium-review.googlesource.com/c/chromium/src/+/3552639. It
was reverted because of flakiness that was caused by not consistent
window position and sizing:
https://chromium-review.googlesource.com/c/chromium/src/+/3553579. It's
fixed in this CL by setting not only the size of the window but the
position too.

Bug: 1286276
Change-Id: I2c30d5dfa68aad597b6945a6c45c307b2bd7995f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3555688
Reviewed-by: Theodore Olsauskas-Warren <sauski@google.com>
Commit-Queue: Olesia Marukhno <olesiamarukhno@google.com>
Cr-Commit-Position: refs/heads/main@{#986388}
NOKEYCHECK=True
GitOrigin-RevId: 4f74096a21ee61c64b064cbd69870dba06ee68ab