Allow a #[gtest] test to specify a C++ class for its TestSuite.
By default the testing::Test class is used, which is equivalent to the
C++ TEST() macro. But just as in C++, with the TEST_F() macro, we need
a way to specify a different class to be the TestSuite. The only
requirement is that the test is a subclass of testing::Test.

We generalize the Gtest TestSuite factory code with a templated
function, but Rust can not call templated C++ functions yet. So we
still wrap that function with the pre-existing
rust_gtest_default_factory() function.

This would allow easily writing other factory functions in C++ targets
outside of //testing on top of the template helper, or even generating
such functions in the future.

We augment the #[gtest] macro to look for a further #[gtest_suite]
macro that names a C++ factory for the TestSuite class. The syntax is:

  #[gtest_suite(some_factory_function)]

Where the factory function is an unsafe extern "C" function pointer that
receives a function pointer and returns a
rust_gtest_interop::GtestFactoryPtr, making its signature:

  unsafe extern "C" fn(f: extern "C" fn()) -> GTestFactoryPtr

In the process, we drop a level of abstraction, so we no longer have a
factory-factory, and Rust works with a function pointer to the C++
function directly.

R=lukasza@chromium.org

Bug: 1305396
Change-Id: I1fe5efa8090435133054b3b7e71f28184b9d4733
Cq-Include-Trybots: luci.chromium.try:android-rust-arm-rel,android-rust-arm-dbg,linux-rust-x64-dbg,linux-rust-x64-rel
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3560902
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#987906}
NOKEYCHECK=True
GitOrigin-RevId: 8029f802d9e66afc5cfd5476d110b9217acf9288