[reland][privacy sandbox settings] Handle notice dismiss in web UI
Dismissing the notice was handled in the dialog view. The action
didn't go through the handler pipeline and didn't trigger HaTS. It also
recorded kNoticeClosedNoAction as well as kNoticeDismiss.

Handle pressing ESC in the web UI instead to go through the same path as
other user actions.

CL was reverted because new interactive tests are flaky on ChromeOS and
Wayland:
https://chromium-review.googlesource.com/c/chromium/src/+/3584774
Disable them on ChromeOS and Wayland for now.

Bug: 1312042
Change-Id: Ib9f59ca83ea2efcd3d0ec2891378690e396a1935
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3581996
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Reviewed-by: Rainhard Findling <rainhard@chromium.org>
Commit-Queue: Olesia Marukhno <olesiamarukhno@google.com>
Cr-Commit-Position: refs/heads/main@{#992472}
NOKEYCHECK=True
GitOrigin-RevId: 1ada708bedb355742054f7c848ba0c1359a61fb3