[infra] Create test specs for linux-blink-rel-dummy try builders.
The linux-blink-rel-dummy builder does not actually exist. Mirroring
non-existent builders is not supported with src-side configs, so in
order to migrate the config for linux-blink-rel, it needs to be made
into a standalone try builder. v8_linux_blink_rel won't be able to share
the config with linux-blink-rel since it won't be able to mirror
linux-blink-rel in order to read its source side spec, so it is also
being made into a standalone try builder.

Bug: 1317387
Change-Id: Id4b0944676869f081ebc5c5018f717fc8afc67e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3591732
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#993479}
NOKEYCHECK=True
GitOrigin-RevId: b989592bf3f3f355a84a02a15235baee8a4352d1