[reland][privacy sandbox settings] Handle notice dismiss in web UI
Dismissing the notice was handled in the dialog view. The action
didn't go through the handler pipeline and didn't trigger HaTS. It also
recorded kNoticeClosedNoAction as well as kNoticeDismiss.

Handle pressing ESC in the web UI instead to go through the same path as
other user actions.

CL was reverted because new interactive tests are flaky on ChromeOS and
Wayland:
https://chromium-review.googlesource.com/c/chromium/src/+/3587187
Disable them on ChromeOS and Wayland for now.

Bug: 1312042
Change-Id: I05b96759284f845c29022fa31e4b51f5e82a125a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3593256
Reviewed-by: Theodore Olsauskas-Warren <sauski@google.com>
Commit-Queue: Olesia Marukhno <olesiamarukhno@google.com>
Cr-Commit-Position: refs/heads/main@{#994088}
NOKEYCHECK=True
GitOrigin-RevId: 9d52ffa06faed331666e4bcbbc0bae3ec8a3e525