Migrate ToTLinuxPGO's tests to bionic
This was overlooked during bionic migration of the clang waterfall
since this bot doesn't specify an OS to run its tests on and instead
uses chromium_swarming recipe_module's default.

Will try changing that default in another CL. But that might lead to
issues, so this pins this particular bot to bionic for safety.

Bug: 1199405
Change-Id: I2949a116904a8d0b1c381dc45a63d286a8ba8e8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3597478
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994340}
NOKEYCHECK=True
GitOrigin-RevId: c9c773ebaa05443d9ef3f0669cc5851e21ec6a67