Add fieldtrial config for AllowHistoryDeletionForChildAccounts
This adds the fieldtrial config, in preparation for starting to enable
this on beta.

I've updated the one test which fails as a result, splitting it into two
for now so that we have coverage of both cases (I've added a TODO to
remove the disabled case as part of the cleanup bug).

Bug: 1318343
Change-Id: I19735a09bf1621e5202071734e5e41ac16e5cef6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3599070
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: James Lee <ljjlee@google.com>
Auto-Submit: James Lee <ljjlee@google.com>
Cr-Commit-Position: refs/heads/main@{#995089}
NOKEYCHECK=True
GitOrigin-RevId: df5e491f81b998745ffe013b53649817411b02a7