[AutoUpdate][OOBE]: Inform EU users about opt out option
Show an additional step during UpdateScreen in OOBE for users from EU when the `kConsumerAutoUpdateToggleAllowed` feature is enabled and update is not needed.

Slightly modify strings of one of the slides in the carousel on the screen, when an update happens.

In the update screen we have a short timer that prevents the screen from being shown if the check for an update happens fast. For EU users that is not the case, as we should show an additional info to them if there is no update.

Bug: 1318136,1161276,1312734
Change-Id: Icc10ea2124e99831024c47c99086862694c1ada6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3602863
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>
Commit-Queue: Danila Kuzmin <dkuzmin@google.com>
Reviewed-by: Alexander Alekseev <alemate@chromium.org>
Cr-Commit-Position: refs/heads/main@{#996679}
NOKEYCHECK=True
GitOrigin-RevId: 9612e5fe073c02bebdc1a8112b6152a2ef770978