Add android-x86 and android 10 tester for WebView CTS tests
We now have working Android Webview CTS tests for Android 10.
We are adding an Android x86 builder and an Android 10 tester
that will rely on this to run Webview CTS tests.

We will remove the rules disabling tree closing and sherrifs not being
notified after confirming this is running correctly.
We will disable the FYI tests at that time.

Test: We have confirmed that the android-10-x86-fyi-rel bot is now
green so these tests won't block the waterfall when they are introduced.

Bug: 1250464
Change-Id: I597c17f79f5ad3d7dbc656f884f76ebdecfa8d9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3613203
Commit-Queue: Rupert Wiser <bewise@chromium.org>
Reviewed-by: Erik Staab <estaab@chromium.org>
Cr-Commit-Position: refs/heads/main@{#999309}
NOKEYCHECK=True
GitOrigin-RevId: 1d0e11d97a592bf281a97111d016e8c5afff5e16