Fix arm64 Android monochrome trybots timing out
The trybots that rely on the Android arm64 Builder (dbg)
are timing out because this builder contains an "all" compile
target. The Android toolchain contains nonstandard builders
for gn meaning "gn analyze" can't normally filter out targets
for us. This means that when our try bots try to execute,
they end up trying to compile everything which results in
5+ hour compile times.

We still want the compile all target to verify that everything
still builds on waterfall.

To fix this, we are removing the all compile target
which slows down our trybots, and we are instead
adding a new builder which will only be added
for compiling all.

Bug: 1246468
Change-Id: Idc5f38fb03a7fb08043be1d43e8aaf0414259318
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3627593
Commit-Queue: Rupert Wiser <bewise@chromium.org>
Reviewed-by: Erik Staab <estaab@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1000052}
NOKEYCHECK=True
GitOrigin-RevId: ecf2346daf4db020dc2927d7bbbd19ba209613cc