Apply GPU's test name substitution hack to more tests
crrev.com/c/3627855 applied the test name substitution to some tests
that use variants, but not all of them.

That's because a test config can be identified by either the key to
its dict in test_suites.pyl, or the value of the 'name' field inside
the dict if it's present.

The initial CL applied the name hack only to configs with the 'name'
field. This will apply it to both.

Bug: 1317188
Change-Id: Ic0f4e7062bb0dc1cfe4e9501128e7025512652fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3631221
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1000106}
NOKEYCHECK=True
GitOrigin-RevId: 5ff45d893419a7236740b298be15691fd93193ac