Intents: Add basic layout changes for Intent Picker grid view
We are adding a new design for the Intent Picker bubble, as part of the
ongoing Link Capturing redesign. The most prominent change will be the
introduction of a grid-based UI for selecting the app to open, instead
of the current list UI.

This CL makes all the necessary changes to the bubble other than the
change from list to grid. In particular,

- The dialog title is changed to 'Open in app'
- There's a new subtitle string at the top of the dialog
- There's no Stay in Chrome option
- The separator between the list and the bottom of the dialog is removed

All of these changes are behind a feature flag (LinkCapturingUiUpdate).
This also includes a separate pixel test, so we can verify that the
flagged changes are correct and the current ToT experience is unchanged.

Bug: 1321501
Change-Id: I97e710d6e9c8cd0880958ee254e733936afcca68
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3640907
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Tim Sergeant <tsergeant@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1002976}
NOKEYCHECK=True
GitOrigin-RevId: f30c250fe4ad9f3b7d5a28e390faaaf3354b57e9