Add finch config for ReportFCPOnlyOnSuccessfulCommit
This adds an entry for the ReportFCPOnlyOnSuccessfulCommit
experiment to field_trial_testing_config.json.

The trial also exposes a flake in the Test PageLoadMetricsBrowserTestWithAnimatedLCPFlag.PageLCPAnimatedImage test, so a timeout is extended to deflake it. (A better fix will be in a follow-up CL)

Bug: 1257607
Change-Id: I3dc8dccb4eb14d73e22896b9c6e372d64f6025d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3636343
Auto-Submit: Ian Clelland <iclelland@chromium.org>
Commit-Queue: Ian Clelland <iclelland@chromium.org>
Reviewed-by: Annie Sullivan <sullivan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1008206}
NOKEYCHECK=True
GitOrigin-RevId: d10689e8dc7056981ed575af31146e963e186a96