Add ExtraTimeMin params to fieldtrial_testing_config.json
This change updates feature params for the DnsHttpsSvcbSchemeUpgrade
experiment.

Besides adding ExtraTimeMin params, it also replaces the feature params
that were deprecated in https://crrev.com/c/3569573.

We are seeing a lot of TimedOut values in the DnsRcode histograms. We
think that the percent-based timeout logic may be too harsh when the
address queries complete very quickly. If that assumption is true, we
should be able to prevent these timeouts by setting the "extra time min"
params.

Change-Id: Ib8eba10c96e76c7ee94f2a39ae84a3d8258d56ca
Fixed: 1330689
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3679943
Reviewed-by: David Benjamin <davidben@chromium.org>
Commit-Queue: Dan McArdle <dmcardle@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1009778}
NOKEYCHECK=True
GitOrigin-RevId: 36ee637b812ee8050d0a0c896d8dc1609dd8b4b1