[flakiness] Explicitly disable Flake Endorser for blink_web_tests
Flake Endorser is implicitly not enabled due to the large size of the
"blink_web_tests" suite. However, some test configs run a subset of
tests of it, so we need to explicitly disable "check_flakiness_for_new_tests" in all test configs of it until the
suite is claimed supported.

Bug: 1335990,1327314
Change-Id: Ie5d73993ad4d0d222137daf18221928f261ee526
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3703541
Reviewed-by: Jeff Yoon <jeffyoon@google.com>
Commit-Queue: Zhaoyang Li <zhaoyangli@chromium.org>
Reviewed-by: Sven Zheng <svenzheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1013733}
NOKEYCHECK=True
GitOrigin-RevId: 07401ab6562c67e3fe608d029e9b580c03e3bd6f