Add HTMLParamElementUrlSupport to fieldtrial_testing_config, +bug fix
This *disables* HTMLParamElementUrlSupport in the
fieldtrial_testing_config, to support a Finch experiment that is slowly
attempting to disable this feature/behavior.

In disabling this for testing, a bug was discovered and fixed. The
previous code skipped handling of attributes on the <object> element
itself, causing issues.

Based on the discovery of this bug, I'll postpone the Finch move to
stable until the version this CL lands within.

Bug: 1315717
Change-Id: Ie869d58dfdcc66f5cfb42d160b9cd97357b0bbf3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3704632
Commit-Queue: Mason Freed <masonf@chromium.org>
Auto-Submit: Mason Freed <masonf@chromium.org>
Reviewed-by: David Baron <dbaron@chromium.org>
Commit-Queue: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1014596}
NOKEYCHECK=True
GitOrigin-RevId: 5d152414cfa22c2361a7f33618b855cfe99c43e5