Remove validating command decoder tests from Mac bots
The passthrough command decoder is enabled by default on Mac. At this
point this is not going to be reverted, so we can stop testing old
paths.
This CL removes validating tests from all Mac bots. In the case of the
optional Intel tests, this CL switches from validating to passthrough
for Webgl2 tests. It switches Webgl2 to passthrough on NVIDIA. It also
switches V8 pixel tests to passthrough.

Bug: chromium:982292
Bug: chromium:1236796
Change-Id: I20df1f7e5e605dfaec97aeaad7ae325520940c7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3704175
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Commit-Queue: Kenneth Russell <kbr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1015076}
NOKEYCHECK=True
GitOrigin-RevId: ef00221a193508737b68e836fa1ce7c28bf58b99