Specify validating command decoder for Android testing.
Now that there is a fieldtrial testing config for Android, we need to
specify that we want to run all tests on validating (or passthrough)
instead of defaulting to whatever is chosen by fieldtrial_testing. For
now, keep running things on validating.

Bug: chromium:1338266
Change-Id: Ic9405c54884318b2564e1c55bc840abbcefa915b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3715354
Commit-Queue: Yuly Novikov <ynovikov@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1017731}
NOKEYCHECK=True
GitOrigin-RevId: 4ff0a31f208ca7ca3dac429fe93a3a54fc04a39e