Give zhiyuans OWNERShip of android testing specs in //testing/buildbot
James is doing a bunch of test/bot config changes as part of tasks
like crbug.com/1303439. They're familiar with the configs and the make-up of the test pools and should be able to be trusted with the contents of these files.

Bug: None
Change-Id: Id7ed08dc8a5bb764d66769230fad5cd74c355b1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3751868
Reviewed-by: Erik Staab <estaab@chromium.org>
Reviewed-by: Gary Tong <gatong@chromium.org>
Commit-Queue: Gary Tong <gatong@chromium.org>
Auto-Submit: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1023428}
NOKEYCHECK=True
GitOrigin-RevId: 9cdafd91d3921b7c8807f9728e32f421a1efc4ab